meshtastic-dashboard/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
paho-mqtt>=1.6.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
snowflake-connector-python>=3.0.0
requests>=2.28.0
python-dotenv>=1.0.0
